Shift hole2 docs to module level docstring (#3579)
Part of #2786
