+++ package/AUTHORS
@@ -186,6 +186,7 @@ Chronological list of authors
   - Meghan Osato
   - Anirvinya G
   - Rishabh Shukla  
+  - Manish Kumar
 
 External code
 -------------
