+++ package/MDAnalysis/analysis/hole2/templates.py
@@ -22,7 +22,7 @@
 #
 
 """HOLE Analysis --- :mod:`MDAnalysis.analysis.hole2.templates`
-=====================================================================================
+===============================================================
 
 :Author: Lily Wang
 :Year: 2020
